backend/requirements.txt
fastapi==0.115.4
sqlmodel==0.0.22
pydantic==2.10.3
pydantic-settings==2.7.1
PyJWT==2.10.1
passlib==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
psycopg2-binary==2.9.10
uvicorn==0.32.1
alembic==1.14.0
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
bcrypt==4.2.1